Change Report type from double to decimal.

+++ src/Events/ProgressReportedEvent.cs
@@ -18,7 +18,7 @@ namespace HierarchicalProgress.Events
         /// <param name="change">The manner in which the report has affected the progress.</param>
         /// <param name="currentProgress">Represents the current progress value in the range of <see cref="ProgressBoundaries"/>.</param>
         /// <param name="progressBoundaries">Represents the minimum and maximum progress values allowed to be reported.</param>
-        public ProgressReportedEventArgs(TProgressValue? previousProgress, TProgressValue reportedProgress, ProgressChange change, Index<double> currentProgress, Range<double> progressBoundaries)
+        public ProgressReportedEventArgs(TProgressValue? previousProgress, TProgressValue reportedProgress, ProgressChange change, Index<decimal> currentProgress, Range<decimal> progressBoundaries)
             : base(previousProgress, reportedProgress)
         {
             Change = change;
@@ -31,10 +31,10 @@ namespace HierarchicalProgress.Events
         
         /// <summary>Represents the current progress value in the range of <see cref="ProgressBoundaries"/>.</summary>
         /// <remarks><see cref="Index{T}.IsFromEnd"/> has to be <see langword="false"/>.</remarks>
-        public Index<double> CurrentProgress { get; set; }
+        public Index<decimal> CurrentProgress { get; set; }
         
         /// <summary>Represents the minimum and maximum progress values allowed to be reported.</summary>
         /// <remarks><see cref="Index{T}.IsFromEnd"/> has to be <see langword="false"/>.</remarks>
-        public Range<double> ProgressBoundaries { get; set; }
+        public Range<decimal> ProgressBoundaries { get; set; }
     }
 }
