+++ src/IHierarchicalProgress.cs
@@ -15,25 +15,25 @@ namespace HierarchicalProgress
     {
         event EventHandler<ProgressResetEventArgs<TProgressReport>> Reset;
 
-        event EventHandler<ProgressCompletedEventArgs<TProgressReport>> Completed; 
-        
+        event EventHandler<ProgressCompletedEventArgs<TProgressReport>> Completed;
+
         event EventHandler<ProgressReportedEventArgs<TProgressReport>> Reported;
-        
+
         /// <summary>Represents minimum and maximum progress value.</summary>
         /// <remarks><see cref="Index{T}.IsFromEnd"/> has to be <see langword="false"/>.</remarks>
-        Range<double> ProgressBoundaries { get; }
-        
+        Range<decimal> ProgressBoundaries { get; }
+
         /// <summary>Represents the current progress value in the range of <see cref="ProgressBoundaries"/>.</summary>
         /// <remarks>
         ///     Any reported progress value will be mapped from <see cref="ReportBoundaries"/> to <see cref="ProgressBoundaries"/> to assign a value to <see cref="Progress"/>.
         ///     <br/>
         ///     <see cref="Index{T}.IsFromEnd"/> has to be <see langword="false"/>.
         /// </remarks>
-        Index<double> Progress { get; }
-        
+        Index<decimal> Progress { get; }
+
         /// <summary>Represents the portion of the <see cref="ProgressBoundaries"/> reserved for slices of the progress.</summary>
-        double AllocatedProgress { get; }
-        
+        decimal AllocatedProgress { get; }
+
         /// <summary>Represents the minimum and maximum progress values allowed to be reported.</summary>
         /// <remarks>
         ///     The upper limit is negatively influenced by <see cref="AllocatedProgress"/> in proportion to the percentage of <see cref="ProgressBoundaries"/>.
@@ -41,22 +41,22 @@ namespace HierarchicalProgress
         ///     <br/>
         ///     <see cref="Index{T}.IsFromEnd"/> has to be <see langword="false"/>.
         /// </remarks>
-        Range<double> ReportBoundaries { get; }
-        
+        Range<decimal> ReportBoundaries { get; }
+
         /// <summary>Indicates whether the progress is completed or not.</summary>
         /// <remarks>Once <see langword="true"/> subsequent progress reports will throw a <see cref="InvalidProgressStateException"/>.</remarks>
         bool IsCompleted { get; }
-        
+
         /// <summary>Represents the latest reported progress.</summary>
         TProgressReport? LatestReport { get; }
-        
+
         /// <summary>Creates a new <see cref="IHierarchicalProgress{TProgressValue}"/>.</summary>
         /// <param name="reportBoundaries">The value assigned to <see cref="ReportProgressBoundaries"/> of the slice.</param>
         /// <param name="allocateProgress">The amount of the progress of this instance to allocate to the slice. Directly added to <see cref="AllocatedProgress"/>.</param>
         /// <returns>A new <see cref="IHierarchicalProgress{TProgressValue}"/> representing a protion of this progress.</returns>
         /// <remarks>Observes changed to the created <see cref="IHierarchicalProgress{TProgressValue}"/> and reports them.</remarks>
-        IHierarchicalProgress<TProgressReport> Slice(Range<double> reportBoundaries, double allocateProgress);
-        
+        IHierarchicalProgress<TProgressReport> Slice(Range<decimal> reportBoundaries, decimal allocateProgress);
+
         /// <summary>Reports the progress provider as complete.</summary>
         /// <param name="report">The report that completes the progress.</param>
         /// <remarks>Once completed subsequent progress reports will throw a <see cref="InvalidProgressStateException"/>.</remarks>
